+++ tensorcircuit/basecircuit.py
@@ -10,7 +10,7 @@ import graphviz
 import tensornetwork as tn
 
 from . import gates
-from .quantum import QuOperator, QuVector
+from .quantum import QuOperator, QuVector, correlation_from_counts, measurement_counts
 from .abstractcircuit import AbstractCircuit
 from .cons import npdtype, backend, dtypestr, contractor, rdtypestr
 from .simplify import _split_two_qubit_gate
@@ -495,7 +495,7 @@ class BaseCircuit(AbstractCircuit):
         self,
         batch: Optional[int] = None,
         allow_state: bool = False,
-        status: Optional[Tensor] = None,
+        random_generator: Optional[Any] = None,
     ) -> Any:
         """
         batched sampling from state or circuit tensor network directly
@@ -505,18 +505,18 @@ class BaseCircuit(AbstractCircuit):
         :param allow_state: if true, we sample from the final state
             if memory allsows, True is prefered, defaults to False
         :type allow_state: bool, optional
-        :param status: random generator,  defaults to None
-        :type status: Optional[Tensor], optional
+        :param random_generator: random generator,  defaults to None
+        :type random_generator: Optional[Any], optional
         :return: List (if batch) of tuple (binary configuration tensor and correponding probability)
         :rtype: Any
         """
         # allow_state = False is compatibility issue
         if not allow_state:
-            if status is None:
-                status = backend.get_random_state()
+            if random_generator is None:
+                random_generator = backend.get_random_state()
 
             if batch is None:
-                seed = backend.stateful_randu(status, shape=[self._nqubits])
+                seed = backend.stateful_randu(random_generator, shape=[self._nqubits])
                 return self.perfect_sampling(seed)
 
             @backend.jit  # type: ignore
@@ -526,7 +526,7 @@ class BaseCircuit(AbstractCircuit):
 
             r = []
 
-            subkey = status
+            subkey = random_generator
             for _ in range(batch):
                 key, subkey = backend.random_split(subkey)
                 r.append(perfect_sampling(key))
@@ -541,12 +541,12 @@ class BaseCircuit(AbstractCircuit):
         if self.is_dm is False:
             p = backend.abs(s) ** 2
         else:
-            p = backend.real(backend.diagonal(s))
-        if status is None:
+            p = backend.abs(backend.diagonal(s))
+        if random_generator is None:
             ch = backend.implicit_randc(a=2**self._nqubits, shape=[nbatch], p=p)
         else:
             ch = backend.stateful_randc(
-                status, a=2**self._nqubits, shape=[nbatch], p=p
+                random_generator, a=2**self._nqubits, shape=[nbatch], p=p
             )
         prob = backend.gather1d(p, ch)
         confg = backend.mod(
@@ -559,6 +559,77 @@ class BaseCircuit(AbstractCircuit):
         if batch is None:
             r = r[0]
         return r
+
+    def sample_expectation_ps(
+        self,
+        x: Optional[Sequence[int]] = None,
+        y: Optional[Sequence[int]] = None,
+        z: Optional[Sequence[int]] = None,
+        shots: Optional[int] = None,
+        random_generator: Optional[Any] = None,
+        **kws: Any,
+    ) -> Tensor:
+        """
+        Compute the expectation with given Pauli string with measurement shots numbers
+
+        :Example:
+
+        >>> c = tc.Circuit(2)
+        >>> c.H(0)
+        >>> c.rx(1, theta=np.pi/2)
+        >>> c.sample_expectation_ps(x=[0], y=[1])
+        -0.99999976
+
+        :param x: index for Pauli X, defaults to None
+        :type x: Optional[Sequence[int]], optional
+        :param y: index for Pauli Y, defaults to None
+        :type y: Optional[Sequence[int]], optional
+        :param z: index for Pauli Z, defaults to None
+        :type z: Optional[Sequence[int]], optional
+        :param shots: number of measurement shots, defaults to None, indicating analytical result
+        :type shots: Optional[int], optional
+        :param random_generator: random_generator, defaults to None
+        :type random_general: Optional[Any]
+        :return: [description]
+        :rtype: Tensor
+        """
+        if self.is_dm is False:
+            c = type(self)(self._nqubits, mps_inputs=self.quvector())  # type: ignore
+        else:
+            c = type(self)(self._nqubits, mpo_dminputs=self.get_dm_as_quoperator())  # type: ignore
+        if x is None:
+            x = []
+        if y is None:
+            y = []
+        if z is None:
+            z = []
+        for i in x:
+            c.H(i)  # type: ignore
+        for i in y:
+            c.rx(i, theta=np.pi / 2)  # type: ignore
+        s = c.state()  # type: ignore
+        if c.is_dm is False:
+            p = backend.abs(s) ** 2
+        else:
+            p = backend.abs(backend.diagonal(s))
+        # readout error can be processed here later
+        # TODO(@refraction-ray): explicit management on randomness
+        mc = measurement_counts(
+            p,
+            counts=shots,
+            sparse=False,
+            is_prob=True,
+            random_generator=random_generator,
+            jittable=True,
+        )
+        x = list(x)
+        y = list(y)
+        z = list(z)
+        r = correlation_from_counts(x + y + z, mc)
+        # TODO(@refraction-ray): analytical standard deviation
+        return r
+
+    sexpps = sample_expectation_ps
 
     def replace_inputs(self, inputs: Tensor) -> None:
         """