+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 ## Unreleased
 
+### Added
+
+- Add `sample_expectation_ps` method for `BaseCircuit`, which measure the Pauli string expectation considering measurement shots
+
+- Add alias `expps` for `expectation_ps` and `sexpps` for `sampled_expectation_ps`
+
+- Add `counts_d2s` and `counts_s2d` in quantum module to transform different representation of measurement shots results
+
+- Add vmap enhanced `parameter_shift_grad` in experimental module (API subjects to change)
+
+- Add `parameter_shift.py` script in examples
+
+- Add jit support and external random management for `tc.quantum.measurement_counts`
+
+### Changed
+
+- `rxx`, `ryy`, `rzz` gates now has 1/2 factor before theta consitent with `rx`, `ry`, `rz` gates. (breaking change)
+
+- replace `status` arguments in `sample` method as `random_generator` (new convention: status for 0, 1 uniform randomness and random_generator for random key) (breaking change)
+
 ## 0.3.1
 
 ### Added